Remove module.modulemap from libdispatch-dev pkg
This avoids conflicting modulemaps when libdispatch-dev is used.

THIS IS A WORKAROUND!

The libdispatch-dev pkg provides `dispatchConfig.cmake` which pulls in
`dispatchExports.cmake`. Both of these files use hardcoded paths to the
libdispatch build directory for libdispatch assets (headers, .so files,
etc). This is a problem when a pkg relies on libdispatch-dev because we
want to pull in these dependencies from the recipe-sysroot and/or
recipe-sysroot-native directories instead of the hardcode build
directory.

But due to the hardcode paths inside `dispatch[Config|Exports].cmake`,
a dependent pkg will incorrectly pull in duplicate modulemaps for
libdispatch leading to conflicts.

For now just remove the libdispatch modulemap from the libdispatch-dev
pkg.