Switch from official Swift toolchain to our own
Background:

In Swift 5.3 support for 16-bit floats was added to the Swift stdlib.
- https://github.com/apple/swift-evolution/blob/master/proposals/0277-float16.md

This support depends on low level functions provided by either libgcc
(GNU) or compiler-rt (LLVM) runtimes.

The Swift compiler leverages the LLVM project to translate Swift code to
machine code. Typically this means that the generated code will be
linked against the compiler-rt runtime - since it is part of the LLVM
project. However, the Yocto system uses the GNU suite of tools,
binaries and libraries by default (gcc, ld, glibc, etc.) - including the
libgcc runtime.

As a result, when building Swift code using the clang compiler provided
by the official Swift toolchain, we instuct it to use the `gnueabihf`
ABI and link against the libgcc runtime. Specifying `gnueabihf` as the
ABI instructs LLVM to translate LLVM IR to GNU specific functions /
instructions where appropriate. In particular, LLVM must map runtime
library calls to runtime library functions based on the specified ABI.
In our case this means mapping to libgcc runtime functions.

Note: at this stage of compilation in LLVM is referred to as "Target
Lowering" and it is hardware specific (X86, ARM, etc.) Since we are
targeting ARM our focus is on ARM Target Lowering.

Issue:

During LLVM Target Lowering for ARM targets, LLVM will support 16-bit
floats in part by using the following for code generation:
```
  // In EABI, these functions have an __aeabi_ prefix, but in GNUEABI they have
  // a __gnu_ prefix (which is the default).
  if (Subtarget->isTargetAEABI()) {
    static const struct {
      const RTLIB::Libcall Op;
      const char * const Name;
      const CallingConv::ID CC;
    } LibraryCalls[] = {
      { RTLIB::FPROUND_F32_F16, "__aeabi_f2h", CallingConv::ARM_AAPCS },
      { RTLIB::FPROUND_F64_F16, "__aeabi_d2h", CallingConv::ARM_AAPCS },
      { RTLIB::FPEXT_F16_F32, "__aeabi_h2f", CallingConv::ARM_AAPCS },
    };
```
https://github.com/apple/llvm-project/blob/swift/release/5.3/llvm/lib/Target/ARM/ARMISelLowering.cpp#L687

This is code is _not_ problematic. However, the assumption detailed in
the comment is. It turns out that LLVM does not generate a call to
`__gnu_d2h_ieee` by default but rather `__aeabi_d2h`! The libgcc runtime
defines the former but not the later and the opposite is true for the
compiler-rt runtime.

The end result is that when compiling for ARM AND requiring 16-bit float
support AND using the libgcc runtime, you'll hit an error at runtime due
to a missing definition for `__aeabi_d2h`.

Fix:

I have patched the Swift 5.3 x86_64 Ubuntu toolchain such that it truly
defaults to using `__gnu_d2h_ieee` unless the build targets AEABI, in
which case the above code snippet will correctly override the default.

References:
- https://github.com/apple/swift-evolution/blob/master/proposals/0277-float16.md
- https://clang.llvm.org/docs/LanguageExtensions.html#half-precision-floating-point
- https://gcc.gnu.org/onlinedocs/gcc/Half-Precision.html#Half-Precision
- https://reviews.llvm.org/D12733
- https://github.com/apple/llvm-project/blob/swift/release/5.3/llvm/lib/Target/ARM/ARMISelLowering.cpp#L687