Use std method of adding native paths for Swift
The typical method for appending to $PATH for native tool is to use the
variable EXTRANATIVEPATH. Unfortunately, EXTRANATIVEPATH allows for
additional paths under the STAGING_BINDIR_NATIVE directory only - add
we're installing the swift toolchain under ${STAGING_DIR_NATIVE}/opt.

To work around this detail, the swift-native recipe now also installs
the symlink "swift-tools" under ${STAGING_BINDIR_NATIVE} which points to
${STAGING_DIR_NATIVE}/opt/usr/bin and we and "swift-tools" to
EXTRANATIVEPATH to ensure the symlink is included in $PATH.

This is a more "correct" way to achieve the same result as before and
allows use to avoid needing to (through trial and error) figure out
which tasks within a recipe needed this $PATH modification. Now, so long
as a recipe DEPENDS on swift-native & adds "swift-tools" to
EXTRANATIVEPATH, $PATH will contain the necessary path during all tasks
that need it.