Fix devtool + swift package edit use case
Specify the EXTERNALSRC_BUILD path to be based on EXTERNALSRC (which is
automatically set when using devtool). The default EXTERNALSRC_BUILD
path breaks our building of swift packages because the default path is
not based on the source dir.

With this commit we specify the default build path for swift components
to be base on the external source path when using an external source.